% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_madlib.R
\name{count_blanks.madlib}
\alias{count_blanks.madlib}
\title{Count The Number of Blanks in a Madlib Object}
\usage{
\method{count_blanks}{madlib}(x)
}
\arguments{
\item{x}{A madlib object}
}
\value{
The number of blanks that are available to be filled in x.
}
\description{
This function counts the number of blanks available to be filled
  in a madlib object.
}
