% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_madlib.R
\name{count_blanks}
\alias{count_blanks}
\title{Count The Number of Blanks}
\usage{
count_blanks(x, ...)
}
\arguments{
\item{x}{An object with blank spaces to fill, e.g., a madlib object.}
}
\value{
The number of blanks that are available to be filled.
}
\description{
This is the generic function to count the number of
  blanks that an object has.
}
