% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_madlib.R
\name{fill}
\alias{fill}
\title{Generic Function for Filling a Madlib}
\usage{
fill(x, ...)
}
\arguments{
\item{x}{An object with blanks to fill.}
}
\value{
A formatted string object.
}
\description{
This is the generic function for filling in the blanks in a madlib
  with the appropriate components.
}
