% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madlib_string.R
\name{format_component.madlib_string}
\alias{format_component.madlib_string}
\title{Format String Madlib Component}
\usage{
\method{format_component}{madlib_string}(madlib_string)
}
\arguments{
\item{madlib_string}{An object of class madlib_string.}
}
\value{
A character string.
}
\description{
This function returns a string component formatted for filling
  a madlib blank.
}
