% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_component.R
\name{format_component}
\alias{format_component}
\title{Format Madlib Component}
\usage{
format_component(x, ...)
}
\arguments{
\item{x}{A madlib component object.}
}
\value{
A string, to be used to fill a madlib blank.
}
\description{
This is the generic function to format a madlib component for
  filling a blank
}
