% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_madlib.R
\name{madlib}
\alias{madlib}
\title{Madlib Style Text Formatting}
\usage{
madlib(x)
}
\arguments{
\item{x}{A character string, representing a madlib object. Must have one
or more blanks "<__>" to fill, using the \code{fill} method.}
}
\value{
An object of class "madlib."
}
\description{
Madlib Style Text Formatting
}
