% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madlib_string.R
\name{madlib_string}
\alias{madlib_string}
\title{String Component For Filling Madlib Blanks}
\usage{
madlib_string(x)
}
\arguments{
\item{x}{A character string, to be used as a component to
fill a madlib blank.}
}
\value{
An object of class "madlib_string."
}
\description{
String Component For Filling Madlib Blanks
}
