% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_madlib.R
\name{new_madlib}
\alias{new_madlib}
\title{Internal Constructor for Madlib Object}
\usage{
new_madlib(x = character())
}
\arguments{
\item{x}{A character string representing a madlib object.}
}
\value{
An object of class "madlib."
}
\description{
This is the internal construction function for an object of class "madlib."
}
