% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madlib_string.R
\name{new_madlib_string}
\alias{new_madlib_string}
\title{Internal Constructor for Madlib String Component}
\usage{
new_madlib_string(x = character())
}
\arguments{
\item{x}{A character string, representing a standard string component
to fill a madlib blank.}
}
\value{
An object of class "madlib_string."
}
\description{
This is the internal construction function for an
  object of class "madlib_string."
}
