% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_madlib.R
\name{validate_madlib}
\alias{validate_madlib}
\title{Validate Madlib Object}
\usage{
validate_madlib(x)
}
\arguments{
\item{x}{An object of class "madlib."}
}
\value{
x. Primarily called for the side effect of throwing an error
  message if x is not a valid madlib.
}
\description{
This function checks whether a madlib object is valid.
}
