% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madlib_string.R
\name{validate_madlib_string}
\alias{validate_madlib_string}
\title{Validate Madlib String Component}
\usage{
validate_madlib_string(x)
}
\arguments{
\item{x}{An object of class "madlib_string."}
}
\value{
x. Primarily called for the side effect of throwing an error
  message if x is not a valid madlib_string.
}
\description{
This function checks whether a madlib string component is valid.
}
